% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_website.R
\name{bs4_utilitr}
\alias{bs4_utilitr}
\title{The bs4 Book output format}
\usage{
bs4_utilitr(extra_dependencies = list(), ...)
}
\arguments{
\item{extra_dependencies}{Additional HTML dependencies}

\item{...}{Additional arguments passed to \code{bookdown::\link{bs4_book}()}.}
}
\value{
An R Markdown output format object to be passed to
  \code{rmarkdown::\link{render}()}.
}
\description{
The bs4 Book output format
}
