% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_website.R
\name{html_document}
\alias{html_document}
\title{The utilitR GitBook output format}
\usage{
html_document(extra_dependencies = list(), ...)
}
\arguments{
\item{extra_dependencies}{Additional HTML dependencies}

\item{...}{Additional arguments passed to \code{bookdown::\link{gitbook}()}.}
}
\value{
An R Markdown output format object to be passed to
  \code{rmarkdown::\link{render}()}.
}
\description{
The utilitR GitBook output format
}
