% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_document.R
\name{html_paged}
\alias{html_paged}
\title{HTML paged version for utilitR document}
\usage{
html_paged(
  ...,
  toc = TRUE,
  toc_depth = 2,
  extra_dependencies = NULL,
  pandoc_args = NULL
)
}
\arguments{
\item{...}{Additional arguments passed to [pagedown::html_paged()].}

\item{toc}{\code{TRUE} to include a table of contents in the output}

\item{toc_depth}{Depth of headers to include in table of contents}

\item{extra_dependencies}{Additional function arguments to pass to the
base R Markdown HTML output formatter \code{\link[rmarkdown]{html_document_base}}}

\item{pandoc_args}{Additional command line options to pass to pandoc}
}
\description{
This function presents an output format necessary
 to build a PDF version of the documentation. This
 is a paged HTML document which means an HTML with
 usueal features for PDFs (page number, A4 formatting, etc.)
}
\seealso{
[html_paged](pagedown::html_paged)
 which is used as a customizable
 basis to render R Markdown documents
}
