% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{render_rmd}
\alias{render_rmd}
\title{Render an Rmd file asis}
\usage{
render_rmd(x)
}
\arguments{
\item{x}{Filepath}
}
\value{
The Rmd formatted into a unique string
}
\description{
Render an Rmd file asis
}
