% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitr.R
\docType{package}
\name{utilitr}
\alias{utilitr}
\title{utilitr: A package to build our documentation locally}
\description{
This package is useful if you want to build our documentation locally.
 If you expect a series of functions from the documentation, this
 is not the right place
}
