% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_website.R
\name{utilitr_dependencies}
\alias{utilitr_dependencies}
\title{Import dependencies for utilitR template}
\usage{
utilitr_dependencies(type = c("html", "pdf"), to_list = FALSE)
}
\arguments{
\item{type}{Which format should we handle ?}
}
\description{
Import dependencies for utilitR template
}
