% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_document.R
\name{html_paged}
\alias{html_paged}
\title{HTML paged version for utilitR document}
\usage{
html_paged(..., extra_dependencies = NULL)
}
\arguments{
\item{...}{Arguments passed to
\code{bookdown::\link[bookdown]{html_document2}}.}

\item{extra_dependencies}{Additional HTML dependencies}
}
\description{
This function presents an output format necessary
 to build a PDF version of the documentation. This
 is a paged HTML document which means an HTML with
 usueal features for PDFs (page number, A4 formatting, etc.)
}
\seealso{
[html_paged](pagedown::html_paged)
 which is used as a customizable
 basis to render R Markdown documents
}
