% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_document.R
\name{html_paged}
\alias{html_paged}
\title{`html_paged` version for utilitR}
\usage{
html_paged(..., extra_dependencies = NULL)
}
\arguments{
\item{...}{Arguments passed to
\code{bookdown::\link[bookdown]{html_document2}}.}

\item{extra_dependencies}{Additional HTML dependencies}
}
\description{
`html_paged` version for utilitR
}
