% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_website.R
\name{utilitr_dependencies}
\alias{utilitr_dependencies}
\title{Import dependencies for utilitR template}
\usage{
utilitr_dependencies(type = c("html", "pdf"), model = c("gitbook", "bs4"))
}
\arguments{
\item{type}{Which output format should we handle ?}

\item{model}{What is the output format model? Choice between "gitbook" and "bs4"}

\item{to_list}{Logical indicating whether we want to return result as
htmltools dependency or simple filepaths list}
}
\description{
Import dependencies for utilitR template
}
