% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_website.R
\name{utilitr_dependencies}
\alias{utilitr_dependencies}
\title{Import dependencies for utilitR template}
\usage{
utilitr_dependencies(output = c("gitbook", "bs4", "pagedown"))
}
\arguments{
\item{output}{What is the output format? Choice between "gitbook", "bs4" and "pagedown"}
}
\description{
Import dependencies for utilitR template
}
